reorganize eva.py
