+++ hydromt/stats/__init__.py
@@ -2,3 +2,5 @@
 """HydroMT statistics"""
 
 from .skills import *
+from .extremes import *
+from .design_events import *
