+++ hydromt/stats/extremes.py
@@ -12,8 +12,6 @@ __all__ = [
     "eva_block_maxima",
     "eva_peaks_over_threshold",
     "get_peaks",
-    "get_peak_hydrographs",
-    "get_hyetograph",
     "get_return_value",
     "fit_extremes",
 ]
@@ -25,6 +23,7 @@ _DISTS = {
 }
 
 ## high level methods
+
 
 def eva_idf(
     da: xr.DataArray,
@@ -54,7 +53,9 @@ def eva_idf(
     xr.Dataset
         IDF table
     """
-    assert np.all(np.diff(durations) > 0), "durations should be monotonically increasing"
+    assert np.all(
+        np.diff(durations) > 0
+    ), "durations should be monotonically increasing"
     dt_max = int(durations[-1])
     da_roll = da.rolling(time=dt_max).construct("duration")
     # get mean intensity for each duration and concat into single dataarray
@@ -209,24 +210,26 @@ def get_peaks(
     xr.DataArray
         Timeseries data with only peak values, all other values are set to NaN
     """
-    
+
     assert 0 < qthresh < 1.0, 'Quantile "qthresh" should be between (0,1)'
     if not ev_type.upper() in _DISTS.keys():
         raise ValueError(
             f"Unknown ev_type {ev_type.upper()}, select from {_DISTS.keys()}."
         )
     bins = None
-    if period in ['year', 'quarter', 'month']:
-        bins = getattr(da['time'].dt, period).values
+    if period in ["year", "quarter", "month"]:
+        bins = getattr(da["time"].dt, period).values
         nperiods = np.unique(bins).size  # FIXME
     else:
-        tstart = da.resample(time=period, label='left').first()['time']
+        tstart = da.resample(time=period, label="left").first()["time"]
         bins = tstart.reindex_like(da, method="ffill").values.astype(float)
         nperiods = np.unique(bins).size
     if ev_type.upper() != "BM":
         bins = None
-        min_sample_size=0
-    func = lambda x: local_max_1d(x, min_dist=min_dist, bins=bins, min_sample_size=min_sample_size)
+        min_sample_size = 0
+    func = lambda x: local_max_1d(
+        x, min_dist=min_dist, bins=bins, min_sample_size=min_sample_size
+    )
     duck = dask.array if isinstance(da.data, dask.array.Array) else np
     lmax = duck.apply_along_axis(func, da.get_axis_num("time"), da)
     # apply POT threshold
@@ -238,103 +241,6 @@ def get_peaks(
     peaks = peaks.assign_coords({"extremes_rate": da_rate})
     peaks.name = "peaks"
     return peaks
-
-
-def get_peak_hydrographs(
-    da: xr.DataArray, 
-    da_peaks: xr.DataArray, 
-    wdw_size: int, 
-    n_peaks: Optional[int] = None,
-    normalize: bool = True,
-) -> xr.DataArray:
-    """Returns a hydrograph of `wdw_size` length around each peak
-    in `da_peaks` with a max value or 1 at the peak. The mean hydrograph can be derived
-    by applying stastics along the 'peak' output dimension.
-
-    Parameters
-    ----------
-    da : xr.DataArray
-        Timeseries data, must have a regular spaced 'time' dimension.
-    da_peaks : xr.DataArray
-        Timeseries data with only peak values, all other values are set to NaN
-    wdw_size : int
-        Length of hydrographs measured in the time series time step.
-    n_peaks : int, optional
-        N largest peaks to return. If None (default) all peaks are returned.
-    normalize : bool, optional
-        If True (default) return peak hydrographs normalized by peak value.
-
-    Returns
-    -------
-    xr.DataArray
-        Hydrographs with new 'peak' and 'dt' dimensions.
-    """
-    assert da.shape == da_peaks.shape, "da and da_peaks must have identical shape"
-    if da_peaks.dtype != 'bool':
-        da_peaks = np.isfinite(da_peaks)
-    if n_peaks is None:  # n_peaks required for output dimensions
-        n_peaks = da_peaks.sum('time').max().compute().item()
-    # temp method with arguments set
-    def _func(ts, peaks, wdw_size=wdw_size, n_peaks=n_peaks, normalize=normalize):
-        return hydrograph_1d(ts, peaks, wdw_size, n_peaks, normalize)
-    if da.ndim == 1:  # fix case with single dim
-        da = da.expand_dims('index')
-        da_peaks = da_peaks.expand_dims('index')
-    da_shape = xr.apply_ufunc(
-        _func,
-        da.chunk({'time':-1}),
-        da_peaks.chunk({'time':-1}),
-        input_core_dims=[["time"], ["time"]],
-        output_core_dims=[["peak", "dt"]],
-        dask_gufunc_kwargs=dict(output_sizes={"peak": n_peaks, "dt": wdw_size}),
-        vectorize=True,
-        dask="parallelized",
-        output_dtypes=[float],
-    ).rename({'dt': 'time'})
-    # set time coordinate
-    t0 = int(np.ceil(wdw_size / 2))
-    da_shape["time"] = xr.IndexVariable("time", (np.arange(wdw_size) - t0))
-    return da_shape.squeeze()
-
-
-def get_hyetograph(da_idf: xr.DataArray, dt: float, length: int) -> xr.DataArray:
-    """Returns design storm hyetograph based on intensity-frequency-duration (IDF) table.
-
-    The input `da_idf` can be obtained as the output of the :py:meth:`eva_idf`.
-    Note: here we use the precipitation intensitity and not the depth as input!
-
-    Parameters
-    ----------
-    da_idf : xr.DataArray
-        IDF data, should contain a 'duration' dimension
-    dt : float
-        Time-step for output hyetograph, same unit as IDF duration.
-    length : int
-        Number of time-step intervals in design storms.
-
-    Returns
-    -------
-    xr.DataArray
-        Design storm hyetograph
-    """
-
-    durations = da_idf["duration"]
-    assert np.all(np.diff(durations) > 0)
-    assert dt >= durations[0]
-
-    t = np.arange(0, durations[-1] + dt, dt)
-    alt_order = np.append(np.arange(1, length, 2)[::-1], np.arange(0, length, 2))
-    # get cummulative precip depth
-    pdepth = (da_idf * durations).reset_coords(drop=True).rename({"duration": "time"})
-    # interpolate to dt temporal resolution
-    pstep = pdepth.interp(time=t).fillna(0).diff("time") / dt
-    # reorder using alternating blocks method
-    pevent = pstep.isel(time=slice(0, length)).isel(time=alt_order)
-    # set time coordinate
-    t0 = int(np.ceil((length+1) / 2))
-    pevent["time"] = xr.IndexVariable("time", (t[1 : length + 1]-t0))
-    pevent.attrs.update(**da_idf.attrs)
-    return pevent
 
 
 def get_return_value(da_params: xr.DataArray, rps: np.ndarray = _RPS) -> xr.DataArray:
@@ -367,8 +273,8 @@ def get_return_value(da_params: xr.DataArray, rps: np.ndarray = _RPS) -> xr.Data
     else:
         extremes_rate = 1.0
 
-    if da_params.ndim == 1: # fix case of single dim
-        da_params = da_params.expand_dims('index')
+    if da_params.ndim == 1:  # fix case of single dim
+        da_params = da_params.expand_dims("index")
     da_rvs = xr.apply_ufunc(
         _return_values_1d,
         da_params.chunk({"dparams": -1}),
@@ -432,9 +338,8 @@ def fit_extremes(
         # trick to include distribution name
         return np.concatenate([params, [distributions.index(d)]])
 
-    
     if da_peaks.ndim == 1:  # fix case of single dim
-        da_peaks = da_peaks.expand_dims('index')
+        da_peaks = da_peaks.expand_dims("index")
     da_params = xr.apply_ufunc(
         _fitopt_1d,
         da_peaks.chunk({"time": -1}),
@@ -454,7 +359,7 @@ def fit_extremes(
     dims = list([d for d in da_params.dims if d != "dparams"])
     coords = dict(
         dparams=xr.IndexVariable("dparams", ["shape", "loc", "scale"]),
-        distribution=xr.IndexVariable(dims=dims, data=distributions)
+        distribution=xr.IndexVariable(dims=dims, data=distributions),
     )
     if "extremes_rate" in da_peaks:  # keep extremes_rate meta data
         coords.update(extremes_rate=da_peaks["extremes_rate"])
@@ -464,59 +369,13 @@ def fit_extremes(
 
 ## PEAKS
 
-@njit
-def hydrograph_1d(
-    ts: np.ndarray, 
-    peaks: np.ndarray, 
-    wdw_size: int, 
-    n_peaks: Optional[int] = None, 
-    normalize: bool = True,
-) -> np.ndarray:
-    """Returns 2D array of shape (`n_peaks`, `wdw_size`) with normalized hydrographs 
-    from time series `ts`
-
-    Parameters
-    ----------
-    ts : np.ndarray of float
-        1D array with constant spaced time series
-    peaks : np.ndarray of bool
-        1D array with constant spaced time series, True where peaks
-    wdw_size : int
-        Size of hydrograph in unit of time series time step
-    n_peaks : int, optional
-        N largest peaks to return. If None (default) all peaks are returned.
-    normalize : bool, optional
-        If True (default) return peak hydrographs normalized by peak value.
-
-    Returns
-    -------
-    np.ndarray
-        normalized hydrographs
-    """
-    assert ts.shape == peaks.shape, "the shapes of ts and peaks mismatch"
-    idxs = np.where(peaks)[0]
-    n0 = idxs.size if n_peaks is None else int(n_peaks)
-    out = np.full((n0, wdw_size), np.nan, ts.dtype)
-    seq = np.argsort(ts[idxs])[::-1]  # sort from large to small
-    n = ts.size
-    d0 = int(np.ceil(wdw_size / 2))
-    d1 = wdw_size - d0
-    for i in range(n0):
-        idx = idxs[seq[i]]
-        idx0 = idx - d0
-        idx1 = idx + d1
-        s = slice(max(0, idx0), min(n + 1, idx1))
-        s1 = slice(max(0, -idx0), n - idx0 if idx1 > n else idx1 - idx0)
-        out[i, s1] = ts[s] / ts[idx] if normalize else ts[s]
-    return out
-
 
 @njit
 def local_max_1d(
-    arr: np.ndarray, 
-    bins: np.ndarray=None, 
-    min_dist: int=0, 
-    min_sample_size: int=0
+    arr: np.ndarray,
+    bins: np.ndarray = None,
+    min_dist: int = 0,
+    min_sample_size: int = 0,
 ) -> np.ndarray:
     """returns boolean index of local maxima in `arr` which are `min_dist` apart
 
@@ -525,8 +384,8 @@ def local_max_1d(
     arr : np.ndarray
         1D time series
     bins : np.ndarray, optional
-        1D array of with uniquely numbered bins (blocks), by default None. 
-        If provided only the largest peak per block is flagged. 
+        1D array of with uniquely numbered bins (blocks), by default None.
+        If provided only the largest peak per block is flagged.
     min_dist : int, optional
         Minimum distance between peaks, by default 0
     min_sample_size : int, optional
@@ -551,8 +410,8 @@ def local_max_1d(
             continue
         dd = i - 1 - imax  # distance to previous peak
         if (imax > 0) and (
-            (bins is None and dd == (min_dist + 1)) or
-            (bins is not None and bins[i-1] != bins[imax] and dd > min_dist)
+            (bins is None and dd == (min_dist + 1))
+            or (bins is not None and bins[i - 1] != bins[imax] and dd > min_dist)
         ):
             if bsize >= min_sample_size:
                 out[imax] = True
@@ -611,7 +470,7 @@ def _aic(x, params, distribution):
 
 
 def _aicc(x, params, distribution):
-    """Return Akaike Information Criterion with correction for small sample size 
+    """Return Akaike Information Criterion with correction for small sample size
     for a frozen distribution"""
     k = len(params)
     aic = _aic(x, params, distribution)
@@ -638,7 +497,7 @@ def _get_return_values(params, distribution, rps=_RPS, extremes_rate=1.0):
 def _get_return_periods(x, a=0.0, extremes_rate=1.0):
     assert np.all(np.isfinite(x))
     b = 1.0 - 2.0 * a
-    ranks = (len(x)+1) - stats.rankdata(x, method="average")
+    ranks = (len(x) + 1) - stats.rankdata(x, method="average")
     freq = ((ranks - a) / (len(x) + b)) * extremes_rate
     rps = 1 / freq
     return rps
@@ -665,6 +524,7 @@ def lmoment_ci(x, distribution, nsample=1000, alpha=0.9, rps=_RPS, extremes_rate
 
 ## PLOTS
 
+
 def plot_return_values(
     x,
     params,
@@ -681,10 +541,13 @@ def plot_return_values(
     import matplotlib.pyplot as plt
 
     rvs_obs = np.sort(x[np.isfinite(x)])
-    params = params[-2:] if distribution == 'gumb' else params
-    rvs_sim = _get_return_values(
-        params, distribution, rps=rps, extremes_rate=extremes_rate
-    )
+    if distribution != "emperical":
+        params = params[-2:] if distribution == "gumb" else params
+        rvs_sim = _get_return_values(
+            params, distribution, rps=rps, extremes_rate=extremes_rate
+        )
+    else:
+        rvs_sim = rvs_obs
     rps_obs = _get_return_periods(rvs_obs, a=a, extremes_rate=extremes_rate)
 
     if x_scale == "gumbel":
@@ -697,9 +560,13 @@ def plot_return_values(
     if ax is None:
         _, ax = plt.subplots(1, 1, figsize=(10, 6))
     ax.plot(xobs, rvs_obs, color=color, marker="o", label=f"plot position", lw=0)
-    ax.plot(xsim, rvs_sim, color=color, ls="--", label=f"{distribution.upper()} fit")
 
-    if alpha is not None and nsample > 0:
+    if distribution != "emperical":
+        ax.plot(
+            xsim, rvs_sim, color=color, ls="--", label=f"{distribution.upper()} fit"
+        )
+
+    if alpha is not None and nsample > 0 and distribution != "emperical":
         urvs = lmoment_ci(
             x,
             distribution,
@@ -758,7 +625,7 @@ def lmoment_fitopt(x, distributions=["gumb", "gev"], criterium="AIC"):
         iterable of distribution names
     criterium: {'AIC', 'AICc', 'BIC'}
         distrition selection criterium
-    
+
     Returns
     -------
     params: ndarray of float
@@ -842,7 +709,7 @@ def _lmomentfit(lmom, distribution):
     # derive parameters for selected distribution
     if distribution in ["gev", "genextreme"]:
         c1 = 2.0 / (3.0 + tau3) - np.log(2.0) / np.log(3.0)
-        k1 = 7.859 * c1 + 2.9554 * (c1 ** 2.0)
+        k1 = 7.859 * c1 + 2.9554 * (c1**2.0)
         s1 = (lmom[1] * k1) / ((1.0 - 2.0 ** (-k1)) * math.gamma(1.0 + k1))
         m1 = lmom[0] - (s1 / k1) * (1.0 - math.gamma(1.0 + k1))
         params = (k1, m1, s1)
@@ -882,7 +749,6 @@ def legendre_shift_poly(n):
     elif n == 1:
         pk = [2, -1]
     else:
-
         pkm2 = np.zeros(n + 1)
         pkm2[-1] = 1
         pkm1 = np.zeros(n + 1)
